+++ src/useInterval/typings.ts
@@ -1,5 +1,5 @@
-export type Cleanup = () => void | null;
-export type Callback = () => Cleanup | void;
+export type Cleanup = void | (() => void);
+export type Callback = () => Cleanup;
 export type Delay = null | number;
 export interface UseIntervalResult {
   resetInterval: () => void;