+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 
 - Add hook `useToggle`
 
+## 0.10.2
+
+- Rename `useResize` to `useWindowResize`
+- No restart timeout in `useTimeout` when change state
+- Fix the reset/repeat timeout in `useIdle`
+
 ## 0.10.0
 
 - Add hook `useIdle`